Cleanup of generic resolver code
- Move `Resolver` and its blanket implementations to where they are
  needed: the introspection submodule.
- Remove the generic parameter from `Resolver` (and fix it to `Value`).
