+++ payas-server-core/src/graphql/execution/resolver.rs
@@ -1,27 +1,12 @@
-use std::iter::FromIterator;
-
-use async_graphql_parser::Positioned;
 use async_trait::async_trait;
 use futures::StreamExt;
-use serde_json::Value;
 
 use crate::graphql::{
-    execution_error::ExecutionError,
-    request_context::{self, RequestContext},
+    execution_error::ExecutionError, request_context::RequestContext,
     validation::field::ValidatedField,
 };
 
 use super::system_context::SystemContext;
-
-#[async_trait]
-pub trait Resolver<R> {
-    async fn resolve_value<'e>(
-        &self,
-        fields: &'e [ValidatedField],
-        system_context: &'e SystemContext,
-        request_context: &'e RequestContext<'e>,
-    ) -> Result<R, ExecutionError>;
-}
 
 #[async_trait]
 pub trait FieldResolver<R>
@@ -44,7 +29,7 @@ where
         &self,
         fields: &[ValidatedField],
         system_context: &SystemContext,
-        request_context: &request_context::RequestContext<'_>,
+        request_context: &RequestContext<'_>,
     ) -> Result<Vec<(String, R)>, ExecutionError> {
         futures::stream::iter(fields.iter())
             .then(|field| async {
@@ -64,88 +49,10 @@ where
 // where
 //     T: FieldResolver<Value>,
 // {
-//     fn resolve_field(&self, system_context: &QueryContext<'_>, field: &Field) -> Value {
+//     fn resolve_field(&self, field: &ValidatedField, system_context: &SystemContext, request_context: &request_context::RequestContext<'_>,) -> Value {
 //         match self {
 //             Some(td) => td.resolve_field(system_context, field),
 //             None => Value::Null,
 //         }
 //     }
 // }
-
-#[async_trait]
-impl<T> Resolver<Value> for T
-where
-    T: FieldResolver<Value> + std::fmt::Debug + Send + Sync,
-{
-    async fn resolve_value<'e>(
-        &self,
-        fields: &'e [ValidatedField],
-        system_context: &'e SystemContext,
-        request_context: &'e request_context::RequestContext<'e>,
-    ) -> Result<Value, ExecutionError> {
-        Ok(Value::Object(FromIterator::from_iter(
-            self.resolve_fields(fields, system_context, request_context)
-                .await?,
-        )))
-    }
-}
-
-#[async_trait]
-impl<T> Resolver<Value> for Option<&T>
-where
-    T: Resolver<Value> + std::fmt::Debug + Send + Sync,
-{
-    async fn resolve_value<'e>(
-        &self,
-        fields: &'e [ValidatedField],
-        system_context: &'e SystemContext,
-        request_context: &'e RequestContext<'e>,
-    ) -> Result<Value, ExecutionError> {
-        match self {
-            Some(elem) => {
-                elem.resolve_value(fields, system_context, request_context)
-                    .await
-            }
-            None => Ok(Value::Null),
-        }
-    }
-}
-
-#[async_trait]
-impl<T> Resolver<Value> for Positioned<T>
-where
-    T: Resolver<Value> + std::fmt::Debug + Send + Sync,
-{
-    async fn resolve_value<'e>(
-        &self,
-        fields: &'e [ValidatedField],
-        system_context: &'e SystemContext,
-        request_context: &'e RequestContext<'e>,
-    ) -> Result<Value, ExecutionError> {
-        self.node
-            .resolve_value(fields, system_context, request_context)
-            .await
-    }
-}
-
-#[async_trait]
-impl<T> Resolver<Value> for Vec<T>
-where
-    T: Resolver<Value> + std::fmt::Debug + Send + Sync,
-{
-    async fn resolve_value<'e>(
-        &self,
-        fields: &'e [ValidatedField],
-        system_context: &'e SystemContext,
-        request_context: &'e request_context::RequestContext<'e>,
-    ) -> Result<Value, ExecutionError> {
-        let resolved: Vec<_> = futures::stream::iter(self.iter())
-            .then(|elem| elem.resolve_value(fields, system_context, request_context))
-            .collect()
-            .await;
-
-        let resolved: Result<Vec<Value>, ExecutionError> = resolved.into_iter().collect();
-
-        Ok(Value::Array(resolved?))
-    }
-}
