+++ payas-server-core/src/graphql/introspection/resolver/mod.rs
@@ -2,6 +2,7 @@ mod directive_resolver;
 mod enum_value_resolver;
 mod field_resolver;
 mod input_value_resolver;
+mod resolver_support;
 pub mod root_resolver;
 mod schema_resolver;
 mod type_resolver;
