+++ payas-server-core/src/graphql/introspection/resolver/type_resolver.rs
@@ -2,11 +2,13 @@ use async_graphql_parser::types::{BaseType, Type, TypeDefinition};
 use async_trait::async_trait;
 use serde_json::Value;
 
-use crate::graphql::execution::resolver::{FieldResolver, Resolver};
+use crate::graphql::execution::resolver::FieldResolver;
 use crate::graphql::execution_error::ExecutionError;
 use crate::graphql::introspection::definition::type_introspection::TypeDefinitionIntrospection;
 use crate::graphql::request_context::RequestContext;
 use crate::graphql::{execution::system_context::SystemContext, validation::field::ValidatedField};
+
+use super::resolver_support::Resolver;
 
 #[derive(Debug)]
 struct BoxedType<'a> {
